Version/1.4.0 (#44)
* compiling version with dto extracted to its own library

* align version to 1.4

* namespace fix

* namespace fix

* fix build order

* fix build order

* rename folders

* missing references

* repair dll & project references

* adapt doc

* change version numbers

* remove commented sections
