+++ LSG.GenericCrud.Dto/LSG.GenericCrud.Dto.csproj
@@ -4,15 +4,19 @@
     <TargetFramework>netcoreapp2.0</TargetFramework>
     <Version>1.4.0</Version>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageId>LSG.GenericCrud.Dto</PackageId>
     <Authors>lonesomegeek</Authors>
     <Company>lonesomegeek</Company>
+    <Product>LSG.GenericCrud.Dto</Product>
+    <RootNamespace>LSG.GenericCrud.Dto</RootNamespace>
+    <AssemblyName>LSG.GenericCrud.Dto</AssemblyName>
     <Description>This library provides a set of tools to enable CRUD rapidly on any aspnetcore Web APIs.</Description>
-    <PackageTags>generic crud webapi aspnetcore</PackageTags>
+    <PackageTags>generic crud webapi aspnetcore dto</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AutoMapper" Version="6.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.0.0" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.0.0" />
   </ItemGroup>
 
   <ItemGroup>
