+++ LSG.GenericCrud.Tests/Controllers/CrudDtoControllerTests.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using AutoMapper;
 using Bogus;
 using LSG.GenericCrud.Controllers;
+using LSG.GenericCrud.Dto.Controllers;
 using LSG.GenericCrud.Exceptions;
 using LSG.GenericCrud.Repositories;
 using LSG.GenericCrud.Tests.Models;
