+++ LSG.GenericCrud.sln
@@ -7,7 +7,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LSG.GenericCrud", "LSG.Gene
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LSG.GenericCrud.Tests", "LSG.GenericCrud.Tests\LSG.GenericCrud.Tests.csproj", "{0359BCEA-7881-488F-83CC-27C3CE94DC30}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LSG.GenericCrud.Extensions", "LSG.GenericCrud.Extensions\LSG.GenericCrud.Extensions.csproj", "{F343DDBD-9F1A-415C-BE8C-500B2619DD1A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LSG.GenericCrud.Extensions", "LSG.GenericCrud.Extensions\LSG.GenericCrud.Extensions.csproj", "{F343DDBD-9F1A-415C-BE8C-500B2619DD1A}"
+	ProjectSection(ProjectDependencies) = postProject
+		{722007F7-B2C5-40E3-B6A8-57AD5A9745F8} = {722007F7-B2C5-40E3-B6A8-57AD5A9745F8}
+	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LSG.GenericCrud.Dto", "LSG.GenericCrud.Dto\LSG.GenericCrud.Dto.csproj", "{4A41E77C-4ECC-49CA-8C23-8413D94B1FB0}"
+	ProjectSection(ProjectDependencies) = postProject
+		{722007F7-B2C5-40E3-B6A8-57AD5A9745F8} = {722007F7-B2C5-40E3-B6A8-57AD5A9745F8}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +35,10 @@ Global
 		{F343DDBD-9F1A-415C-BE8C-500B2619DD1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F343DDBD-9F1A-415C-BE8C-500B2619DD1A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F343DDBD-9F1A-415C-BE8C-500B2619DD1A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A41E77C-4ECC-49CA-8C23-8413D94B1FB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A41E77C-4ECC-49CA-8C23-8413D94B1FB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A41E77C-4ECC-49CA-8C23-8413D94B1FB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A41E77C-4ECC-49CA-8C23-8413D94B1FB0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
