+++ LSG.GenericCrud/Controllers/HistoricalCrudController.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Linq;
-using AutoMapper;
 using LSG.GenericCrud.Exceptions;
 using LSG.GenericCrud.Models;
 using LSG.GenericCrud.Repositories;
 using Microsoft.AspNetCore.Mvc;
-using Newtonsoft.Json;
 
 namespace LSG.GenericCrud.Controllers
 {
@@ -59,116 +56,6 @@ namespace LSG.GenericCrud.Controllers
             try
             {
                 return Ok(_historicalDal.GetHistory(id));
-            }
-            catch (EntityNotFoundException ex)
-            {
-                return NotFound();
-            }
-        }
-    }
-
-    public class HistoricalCrudController<TDto, TEntity> :
-        Controller
-        where TDto : class, IEntity, new()
-        where TEntity : class, IEntity, new()
-    {
-        private readonly HistoricalCrud<TEntity> _dal;
-        private readonly IMapper _mapper;
-
-        public HistoricalCrudController(HistoricalCrud<TEntity> dal, IMapper mapper)
-        {
-            _dal = dal;
-            _mapper = mapper;
-        }
-
-        [HttpGet]
-        public IActionResult GetAll()
-        {
-            var entities = _dal.GetAll();
-            var dtos = entities.Select(_ => _mapper.Map<TDto>(_));
-            return Ok(dtos);
-        }
-
-        [Route("{id}")]
-        [HttpGet]
-        public IActionResult GetById(Guid id)
-        {
-            try
-            {
-                return Ok(_mapper.Map<TDto>(_dal.GetById(id)));
-            }
-            catch (EntityNotFoundException ex)
-            {
-                return NotFound();
-            }
-        }
-
-        /// <summary>
-        /// Creates the specified entity.
-        /// </summary>
-        /// <param name="entity">The entity.</param>
-        [HttpPost("")]
-        public IActionResult Create([FromBody] TDto dto)
-        {
-            var entity = _mapper.Map<TEntity>(dto);
-            var createdEntity = _dal.Create(entity);
-            return Ok(_mapper.Map<TDto>(createdEntity));
-        }
-
-        [HttpPut("{id}")]
-        public IActionResult Update(Guid id, [FromBody] TDto dto)
-        {
-            try
-            {
-                var entity = _mapper.Map<TEntity>(dto);
-                _dal.Update(id, entity);
-                return Ok();
-            }
-            catch (EntityNotFoundException ex)
-            {
-                return NotFound();
-            }
-        }
-
-        [HttpDelete("{id}")]
-        public IActionResult Delete(Guid id)
-        {
-            try
-            {
-                _dal.Delete(id);
-                return Ok();
-            }
-            catch (EntityNotFoundException ex)
-            {
-                return NotFound();
-            }
-
-        }
-
-        [HttpPost("{entityId}/restore")]
-        public IActionResult Restore(Guid entityId /*, string entityName*/)
-        {
-            try
-            {
-                return Ok(_dal.Restore(entityId));
-            }
-            catch (EntityNotFoundException ex)
-            {
-                return NotFound();
-            }
-        }
-
-        /// <summary>
-        /// Gets the history.
-        /// </summary>
-        /// <param name="id">The identifier.</param>
-        /// <returns></returns>
-        [HttpGet("{id}/history")]
-        public IActionResult GetHistory(Guid id)
-        {
-            try
-            {
-                return Ok(_dal.GetHistory(id));
             }
             catch (EntityNotFoundException ex)
             {
