+++ docs/FeatureDTO.md
@@ -13,6 +13,7 @@ First thing to do, is abstracting the business entity tighted to technologies (i
 - ...
 
 I'm not a master in technical explanation of DTO, but it is really working. And my library is supporting it natively, so ... USE IT :)
+You will need to import this nuget package to get DTO feature working: [LSG.GenericCrud.Dto](https://www.nuget.org/packages/LSG.GenericCrud.Dto/)
 
 ## What is happening behind the scenes
 Before getting into further details, I think this is important for you to understand what is happening behind the scenes.
